requirement.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
langchain-community
requests
python-dotenv
duckduckgo-search
fuzzywuzzy
pyautogui
pynput
python-Levenshtein
livekit
SpeechRecognition
PyAudio
numpy<2.0
pyttsx3
mediapipe==0.10.21
ultralytics
Pillow
diffusers
transformers
safetensors
accelerate
playsound==1.2.2
gTTS
google-generativeai
